requirements.txt
streamlit==1.51.0
supabase==2.0.0
python-dotenv==1.0.0
requests==2.31.0

# Data handling (used by Streamlit UI and scripts)
pandas==2.2.3
openpyxl==3.1.5

# Optional integrations (code imports are guarded where needed)
openai==1.3.5
stripe==7.6.0
python-jose==3.3.0
apscheduler==3.10.4
cloudinary==1.41.0
